requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
beautifulsoup4==4.11.1
    # via bs4
bs4==0.0.1
    # via chatgptpy
certifi==2022.12.7
    # via requests
charset-normalizer==2.0.12
    # via requests
chatgptpy==1.0.2
    # via -r requirements.in
colorama==0.4.6
    # via chatgptpy
cssselect2==0.7.0
    # via svglib
greenlet==2.0.1
    # via sqlalchemy
idna==3.4
    # via requests
langchain==0.0.27
    # via -r requirements.in
lxml==4.9.1
    # via svglib
numpy==1.19.5
    # via langchain
pillow==9.3.0
    # via reportlab
pydantic==1.9.2
    # via langchain
python-dotenv==0.20.0
    # via -r requirements.in
pyyaml==6.0
    # via langchain
reportlab==3.6.12
    # via
    #   chatgptpy
    #   svglib
requests==2.27.1
    # via
    #   chatgptpy
    #   langchain
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlalchemy==1.4.44
    # via langchain
svglib==1.4.1
    # via chatgptpy
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
tls-client==0.1.5
    # via chatgptpy
typing-extensions==4.1.1
    # via pydantic
urllib3==1.26.13
    # via requests
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2
